# backend/src/main/java/com/groww/chatbot/exchanges/CategoryResponse.java
package com.groww.chatbot.exchanges;

import lombok.Data;
/**
 * exchanges class
 * for FAQ categories (or subcategories) response
 */


@Data
public class CategoryResponse {

    private String id;
    private String title;

}